% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfboxjars-package.R
\docType{package}
\name{pdfboxjars}
\alias{pdfboxjars}
\alias{pdfboxjars-package}
\title{Java '.jar' Files for 'pdfbox'}
\description{
Java '.jar' Files for 'pdfbox'
}
\author{
Bob Rudis (bob@rud.is)
}
